+++ servers/Azure.Mcp.Server/CHANGELOG.md
@@ -6,6 +6,8 @@ The Azure MCP Server updates automatically by default whenever a new release com
 
 ### Features Added
 
+- Added --tool option to start Azure MCP server with a list of specific tools [[#685](https://github.com/microsoft/mcp/issues/685)]
+
 ### Breaking Changes
 
 ### Bugs Fixed